Refactor: changed for Router
