+++ src/Components/MovieCard/MovieCard.js
@@ -1,20 +1,25 @@
 import './MovieCard.css';
 import CarrotRating from '../CarrotRating/CarrotRating';
 import PropTypes from 'prop-types';
+import { Link } from 'react-router-dom'
 
 function MovieCard({
   title, id, poster_path, average_rating, handleClick }) {
 
   return (
-    <div className="movie-card" onClick={() => handleClick(id)}>
-      <h3 className="movie-title">{title}</h3>
-      <img className="movie-poster" src={poster_path} alt="movie poster">{ }</img>
-      <div className="movie-card-rating">
-        <div>Movie Rating:</div>
-        {/* <div className="rating">{average_rating}</div> */}
-        <CarrotRating rating={average_rating.toFixed(0)} />
+    <Link to={`/movies/${id}`}>
+     {/* <div className="movie-card" onClick={() => handleClick(id)}> */}
+      <div className="movie-card">
+        <h3 className="movie-title">{title}</h3>
+        <img className="movie-poster" src={poster_path} alt="movie poster">{ }</img>
+        <div className="movie-card-rating">
+          <div>Movie Rating:</div>
+          {/* <div className="rating">{average_rating}</div> */}
+          <CarrotRating rating={average_rating.toFixed(0)} />
+        </div>
       </div>
-    </div>)
+    </Link>
+  )
 };
 
 MovieCard.propTypes = {
